+++ .Build/get_version_of_extension.php
@@ -1,5 +1,0 @@
-<?php
-$_EXTKEY = 'yoast_seo';
-include('ext_emconf.php');
-
-echo $EM_CONF[$_EXTKEY]['version'];