+++ Classes/Form/Element/Cornerstone.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 namespace YoastSeoForTypo3\YoastSeo\Form\Element;
 
 use TYPO3\CMS\Backend\Form\AbstractNode;
@@ -27,7 +28,7 @@ class Cornerstone extends AbstractNode
         );
     }
 
-    public function render()
+    public function render(): array
     {
         $resultArray = $this->initializeResultArray();