+++ Classes/Form/Element/RelatedKeyphrases.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 namespace YoastSeoForTypo3\YoastSeo\Form\Element;
 
 use TYPO3\CMS\Backend\Form\AbstractNode;
@@ -15,7 +16,7 @@ class RelatedKeyphrases extends AbstractNode
         parent::__construct($nodeFactory, $data);
     }
 
-    public function render()
+    public function render(): array
     {
         $resultArray = $this->initializeResultArray();