+++ CHANGELOG.md
@@ -8,6 +8,13 @@ We will follow [Semantic Versioning](http://semver.org/).
 ## Yoast SEO Premium for TYPO3
 Besides the free version of our plugin, we also have a premium version. The free version enables you to do all necessary optimizations. With the premium version, we make it even easier to do! More information can be found on https://www.maxserv.com/yoast.
 
+## [UNRELEASED]
+### Changed
+* Dropped CMS8 support, removed all 8-related classes and configuration
+* Introduced strict typing and return-types to all classes
+### Fixed
+* 
+
 ## 7.1.3 December 23, 2020
 ### Fixed
 * Excluded some more unnecessary files and folders from the TER release